requirements.txt
streamlit
yfinance
pandas
numpy
matplotlib
google-generativeai